requirements.txt
Flask
crowd
jira
feedparser
python-dateutil
gunicorn
Flask-Cache
